Split desktop app orchestration
