+++ docs/DECISIONS.md
@@ -13,6 +13,7 @@
 - QuillUI is the UI direction, but core tests must not depend on a dirty local QuillUI checkout.
 - Platform-specific code belongs in adapter packages, not the app target.
 - The first desktop executable is `quill-code-desktop`, built with SwiftUI over the same `WorkspaceSurface` contract used by the HTML/Playwright harness. This keeps native UI work testable before the full QuillUI adapter exists.
+- The desktop executable entry file is `QuillCodeDesktopApp.swift`, not `main.swift`. Swift treats `main.swift` as a top-level script entry in multi-file executable targets, so the native `@main` app lives in a named app file while commands, menu-bar UI, OAuth loopback, notifications, browser fetching, and task coordination stay in focused desktop adapter files.
 - Desktop runtime selection defaults to mock LLM for no-key demos, switches to live TrustedRouter when an environment or stored secret key exists, and supports `QUILLCODE_USE_MOCK_LLM=true` for deterministic test runs.
 - The desktop model picker is data-driven from the TrustedRouter catalog. It keeps Nike 1.0 (`trustedrouter/fast`) first and Prometheus 1.0 (`tr/fusion`) adjacent as the deterministic fallback, groups options by category, and refreshes live catalog data only when an env or stored key exists.
 - TrustedRouter model defaults, provider/model aliases, bundled fallback catalog entries, API base URL, display-name normalization, and sort keys live in `QuillCodeCore`. The live TrustedRouter adapter and app picker both merge sparse catalogs with those bundled defaults, so the default/next-option ordering is consistent across CLI, SwiftUI, and Playwright harness surfaces.
@@ -137,6 +138,7 @@
 - Terminal and browser utility panes live in `QuillCodeTerminalBrowserPaneView.swift`. Terminal history rows, browser navigation, page snapshots, comments, and pane-local draft state should evolve there; `WorkspaceSwiftUIView.swift` should only decide when the panes are visible and route the resulting actions.
 - Secondary utility panes live in `QuillCodeSecondaryPanesView.swift`. Extensions, Memories, Automations, their count/status pills, empty states, and compact cards should evolve there; `WorkspaceSwiftUIView.swift` should only decide pane placement and route commands.
 - Workspace modal surfaces live in `QuillCodeWorkspaceDialogs.swift`. Command palette, keyboard shortcuts, chat search, rename sheets, and worktree create/remove sheets should evolve there; `WorkspaceSwiftUIView.swift` should only decide when a sheet is presented and route completed actions.
+- Desktop app bootstrap lives in `QuillCodeDesktopApp.swift` and should remain declarative. Native commands, menu-bar layout, OAuth loopback capture, browser fetching, automation notifications, and controller orchestration live in separate desktop files. Parity gates intentionally scan the whole `Sources/quill-code-desktop` folder so extraction can continue without pushing implementation details back into scene composition.
 - Workspace command IDs reduce through `WorkspaceCommandPlan` before touching `WorkspaceModel` side effects. The planner owns prefix parsing, draft-prefill mapping, static action mapping, automation quick recurrence parsing, and canonical command-ID-to-tool-name mapping, while `WorkspaceModel` owns mutations and tool dispatch. This keeps command palette rows, slash-command templates, automation row actions, MCP lifecycle actions, memory actions, and git shortcuts from drifting as the command set grows.
 - Automation record construction and due-run draft creation live in `WorkspaceAutomationEngine.swift`, not inside `WorkspaceModel`. `WorkspaceAutomationFactory` owns concrete `QuillAutomation` records and schedule helpers, `WorkspaceAutomationRunner` owns due filtering, recurrence advancement, and generated thread drafts, and `WorkspaceModel` applies those drafts to UI selection, project refresh, stores, and notification reports. This keeps scheduler behavior testable without pulling project persistence or top-bar state into pure automation tests.
 - Multi-agent collaboration should happen through pull requests and a serialized merge train rather than direct pushes to `main`. CI runs on `pull_request`, `push`, and `merge_group` events. The train watches `merge-train` and `automerge` labels, ignores draft or `do-not-merge` PRs, processes only the oldest eligible PR, updates behind branches before merging, and merges only after all non-train checks are successful. This keeps Codex/agent work reviewable while preserving a green mainline.
